add: apply trip detail apis, remove mock data
